+++ tests/GitExtension.spec.tsx
@@ -67,15 +67,7 @@ describe('IGitExtension', () => {
       return Promise.resolve(response);
     });
 
-    const app = {
-      commands: {
-        hasCommand: jest.fn().mockReturnValue(true)
-      },
-      docRegistry: {
-        getFileTypesForPath: jest.fn().mockReturnValue([])
-      }
-    };
-    model = new GitExtension(fakeRoot, app as any, docmanager as any);
+    model = new GitExtension(fakeRoot, docmanager as any);
   });
 
   describe('#pathRepository', () => {
