+++ tests/commands.spec.tsx
@@ -1,18 +1,19 @@
+import { JupyterFrontEnd } from '@jupyterlab/application';
+import { showDialog } from '@jupyterlab/apputils';
+import { CommandRegistry } from '@lumino/commands';
 import 'jest';
+import { addCommands, CommandIDs } from '../src/commandsAndMenu';
 import * as git from '../src/git';
 import { GitExtension } from '../src/model';
-import { IGitExtension } from '../src/tokens';
-
-import { CommandIDs, addCommands } from '../src/commandsAndMenu';
-import { CommandRegistry } from '@lumino/commands';
-import { JupyterFrontEnd } from '@jupyterlab/application';
+import { Git } from '../src/tokens';
 
 jest.mock('../src/git');
+jest.mock('@jupyterlab/apputils');
 
 describe('git-commands', () => {
   const mockGit = git as jest.Mocked<typeof git>;
   let commands: CommandRegistry;
-  let model: IGitExtension;
+  let model: GitExtension;
   let mockResponses: {
     [url: string]: {
       body?: (request: Object) => string;
@@ -74,7 +75,7 @@ describe('git-commands', () => {
       shell: null as any
     };
     model = new GitExtension(app as any);
-    addCommands(app as JupyterFrontEnd, model, null, null);
+    addCommands(app as JupyterFrontEnd, model, null, null, null);
   });
 
   describe('git:add-remote', () => {
@@ -122,6 +123,64 @@ describe('git-commands', () => {
       expect(mockGit.httpGitRequest).toBeCalledWith('/git/remote/add', 'POST', {
         top_repo_path: path,
         url
+      });
+    });
+  });
+
+  describe('git:context-discard', () => {
+    ['staged', 'partially-staged', 'unstaged', 'untracked'].forEach(status => {
+      [' ', 'M', 'A'].forEach(x => {
+        it(`status:${status} - x:${x} may reset and/or checkout`, async () => {
+          const mockDialog = showDialog as jest.MockedFunction<
+            typeof showDialog
+          >;
+          mockDialog.mockResolvedValue({
+            button: {
+              accept: true,
+              caption: '',
+              className: '',
+              displayType: 'default',
+              iconClass: '',
+              iconLabel: '',
+              label: ''
+            },
+            value: undefined
+          });
+          const spyReset = jest.spyOn(model, 'reset');
+          spyReset.mockResolvedValueOnce(undefined);
+          const spyCheckout = jest.spyOn(model, 'checkout');
+          spyCheckout.mockResolvedValueOnce(undefined);
+
+          const path = 'file/path.ext';
+          model.pathRepository = '/path/to/repo';
+          await model.ready;
+
+          await commands.execute(CommandIDs.gitFileDiscard, {
+            x,
+            y: ' ',
+            from: 'from',
+            to: path,
+            status: status as Git.Status,
+            is_binary: false
+          });
+
+          if (status === 'staged' || status === 'partially-staged') {
+            expect(spyReset).toHaveBeenCalledWith(path);
+          } else if (status === 'unstaged') {
+            expect(spyReset).not.toHaveBeenCalled();
+            expect(spyCheckout).toHaveBeenCalledWith({ filename: path });
+          } else if (status === 'partially-staged') {
+            expect(spyReset).toHaveBeenCalledWith(path);
+            if (x !== 'A') {
+              expect(spyCheckout).toHaveBeenCalledWith({ filename: path });
+            } else {
+              expect(spyCheckout).not.toHaveBeenCalled();
+            }
+          }
+
+          spyReset.mockRestore();
+          spyCheckout.mockRestore();
+        });
       });
     });
   });
